allow objects in json postMessage data
